fix: fixed compiler error where fuel data needed to be a long

+++ src/main/java/dev/galacticraft/api/rocket/entity/RequiresFuel.java
@@ -11,9 +11,9 @@ public interface RequiresFuel extends ScalableFuelLevel {
     @Nullable
     Fluid getFuelTankFluid();
 
-    int getFuelTankAmount();
+    long getFuelTankAmount();
 
-    int getFuelTankCapacity();
+    long getFuelTankCapacity();
 
     boolean isTankEmpty();
